refactor: Usecase까지 폴더링
