+++ OwnMyWay/OwnMyWay/Domain/Usecases/DefaultAddRecordUsecase.swift
@@ -7,14 +7,6 @@
 
 import Foundation
 
-protocol AddRecordUsecase {
-    func executeValidationTitle(with title: String?) -> Bool
-    func executeValidationDate(with date: Date?) -> Bool
-    func executeValidationCoordinate(with coordinate: Location) -> Bool
-    func executePickingPhoto(with url: URL, completion: (Result<String, Error>) -> Void)
-    func executeRemovingPhoto(of photoID: String, completion: (Result<Void, Error>) -> Void)
-}
-
 struct DefaultAddRecordUsecase: AddRecordUsecase {
 
     private let repository: TravelRepository
