+++ OwnMyWay/OwnMyWay/Domain/Usecases/DefaultDetailRecordUsecase.swift
@@ -7,13 +7,6 @@
 
 import Foundation
 
-protocol DetailRecordUsecase {
-    func executeRecordUpdate(record: Record, completion: @escaping (Result<Void, Error>) -> Void)
-    func executeRecordDeletion(
-        at record: Record, completion: @escaping (Result<Void, Error>) -> Void
-    )
-}
-
 struct DefaultDetailRecordUsecase: DetailRecordUsecase {
 
     private let repository: TravelRepository
