+++ OwnMyWay/OwnMyWay/Domain/Usecases/DefaultEnterDateUsecase.swift
@@ -7,10 +7,6 @@
 
 import Foundation
 
-protocol EnterDateUsecase {
-    func executeEnteringDate(firstDate: Date, secondDate: Date, completion: ([Date]) -> Void)
-}
-
 struct DefaultEnterDateUsecase: EnterDateUsecase {
     func executeEnteringDate(firstDate: Date, secondDate: Date, completion: ([Date]) -> Void) {
         completion(Array(Set([firstDate, secondDate])).sorted())
