+++ OwnMyWay/OwnMyWay/Domain/Usecases/DefaultReservedTravelUsecase.swift
@@ -7,22 +7,6 @@
 
 import Foundation
 
-protocol ReservedTravelUsecase {
-    func executeDeletion(of travel: Travel, completion: @escaping (Result<Void, Error>) -> Void)
-    func executeLandmarkAddition(
-        of travel: Travel,
-        completion: @escaping (Result<Void, Error>) -> Void
-    )
-    func executeLandmarkDeletion(
-        at landmark: Landmark,
-        completion: @escaping (Result<Void, Error>) -> Void
-    )
-    func executeFlagUpdate(
-        of travel: Travel,
-        completion: @escaping (Result<Travel, Error>) -> Void
-    )
-}
-
 struct DefaultReservedTravelUsecase: ReservedTravelUsecase {
 
     let repository: TravelRepository
