+++ OwnMyWay/OwnMyWay/Infrastructure/Error/EnterTitleError.swift
@@ -1,0 +1,12 @@
+//
+//  EnterTitleError.swift
+//  OwnMyWay
+//
+//  Created by 유한준 on 2021/11/30.
+//
+
+import Foundation
+
+enum EnterTitleError: Error {
+    case nilTitle
+}
