+++ OwnMyWay/OwnMyWay/Presentation/CompleteScene/CompleteCreation/Interface/Usecase/CompleteCreationUsecase.swift
@@ -1,0 +1,12 @@
+//
+//  CompleteCreationUsecase.swift
+//  OwnMyWay
+//
+//  Created by 유한준 on 2021/11/30.
+//
+
+import Foundation
+
+protocol CompleteCreationUsecase {
+    func executeCreation(travel: Travel, completion: @escaping (Result<Void, Error>) -> Void)
+}
