+++ OwnMyWay/OwnMyWay/Presentation/HomeScene/Interface/Usecase/HomeUsecase.swift
@@ -1,0 +1,12 @@
+//
+//  HomeUsecase.swift
+//  OwnMyWay
+//
+//  Created by 유한준 on 2021/11/30.
+//
+
+import Foundation
+
+protocol HomeUsecase {
+    func executeFetch(completion: @escaping (Result<[Travel], Error>) -> Void)
+}
